Moved iOS extension functions in Extensions.kt
