+++ record-core/src/iosMain/kotlin/dev/theolm/record/RecordCore.ios.kt
@@ -2,8 +2,6 @@
 
 package dev.theolm.record
 
-import dev.theolm.record.config.OutputFormat
-import dev.theolm.record.config.OutputLocation
 import dev.theolm.record.config.RecordConfig
 import dev.theolm.record.config.SampleRate
 import dev.theolm.record.error.NoOutputFileException
@@ -30,17 +28,8 @@ import platform.AVFAudio.AVLinearPCMIsFloatKey
 import platform.AVFAudio.AVNumberOfChannelsKey
 import platform.AVFAudio.AVSampleRateKey
 import platform.AVFAudio.setActive
-import platform.CoreAudioTypes.AudioFormatID
-import platform.CoreAudioTypes.kAudioFormatMPEG4AAC
-import platform.CoreAudioTypes.kAudioFormatLinearPCM
-import platform.Foundation.NSDocumentDirectory
 import platform.Foundation.NSError
-import platform.Foundation.NSFileManager
-import platform.Foundation.NSURL
 import platform.Foundation.NSURL.Companion.fileURLWithPath
-import platform.Foundation.NSUserDomainMask
-import platform.Foundation.temporaryDirectory
-import kotlin.time.TimeSource
 
 internal actual object RecordCore {
     private var recorder: AVAudioRecorder? = null
@@ -140,29 +129,5 @@ internal actual object RecordCore {
                 }
             }
         }
-    }
-
-    private fun RecordConfig.getOutput(): String {
-        val timestamp = TimeSource.Monotonic.markNow().toString()
-        val fileName = "${timestamp}${outputFormat.extension}"
-
-        return when (this.outputLocation) {
-            OutputLocation.Cache -> "${NSFileManager.defaultManager.temporaryDirectory.path}/$fileName"
-            OutputLocation.Internal -> {
-                val urls = NSFileManager.defaultManager.URLsForDirectory(
-                    NSDocumentDirectory,
-                    NSUserDomainMask
-                )
-                val documentsURL = urls.first() as? NSURL ?: throw NoOutputFileException()
-                "${documentsURL.path!!}/$fileName"
-            }
-
-            is OutputLocation.Custom -> "${this.outputLocation.path}/$fileName"
-        }
-    }
-
-    private fun OutputFormat.toAVFormatID(): AudioFormatID = when (this) {
-        OutputFormat.MPEG_4 -> kAudioFormatMPEG4AAC
-        OutputFormat.WAV -> kAudioFormatLinearPCM
     }
 }
